Use double rather than float when appropriate to improve precision of heatmaps

+++ src/plugins/heatmap/heatmap.cpp
@@ -107,8 +107,8 @@ void Heatmap::run()
     QgsRectangle myBBox = d.bbox();
     int columns = d.columns();
     int rows = d.rows();
-    float cellsize = d.cellSizeX(); // or d.cellSizeY();  both have the same value
-    float myDecay = d.decayRatio();
+    double cellsize = d.cellSizeX(); // or d.cellSizeY();  both have the same value
+    double myDecay = d.decayRatio();
 
     // Start working on the input vector
     QgsVectorLayer* inputLayer = d.inputVectorLayer();
@@ -164,8 +164,8 @@ void Heatmap::run()
     int wField = 0;
 
     // Handle different radius options
-    float radius;
-    float radiusToMapUnits = 1;
+    double radius;
+    double radiusToMapUnits = 1;
     int myBuffer;
     if ( d.variableRadius() )
     {
@@ -226,7 +226,7 @@ void Heatmap::run()
       // If radius is variable then fetch it and calculate new pixel buffer size
       if ( d.variableRadius() )
       {
-        radius = myFeature.attribute( rField ).toFloat() * radiusToMapUnits;
+        radius = myFeature.attribute( rField ).toDouble() * radiusToMapUnits;
         myBuffer = bufferSize( radius, cellsize );
       }
 
@@ -241,17 +241,17 @@ void Heatmap::run()
       poBand->RasterIO( GF_Read, xPosition, yPosition, blockSize, blockSize,
                         dataBuffer, blockSize, blockSize, GDT_Float32, 0, 0 );
 
-      float weight = 1.0;
+      double weight = 1.0;
       if ( d.weighted() )
       {
-        weight = myFeature.attribute( wField ).toFloat();
+        weight = myFeature.attribute( wField ).toDouble();
       }
 
       for ( int xp = 0; xp <= myBuffer; xp++ )
       {
         for ( int yp = 0; yp <= myBuffer; yp++ )
         {
-          float distance = sqrt( pow( xp, 2.0 ) + pow( yp, 2.0 ) );
+          double distance = sqrt( pow( xp, 2.0 ) + pow( yp, 2.0 ) );
 
           // is pixel outside search bandwidth of feature?
           if ( distance > myBuffer )
@@ -259,7 +259,7 @@ void Heatmap::run()
             continue;
           }
 
-          float pixelValue = weight * ( 1 - (( 1 - myDecay ) * distance / myBuffer ) );
+          double pixelValue = weight * ( 1 - (( 1 - myDecay ) * distance / myBuffer ) );
 
           // clearing anamolies along the axes
           if ( xp == 0 && yp == 0 )
@@ -305,7 +305,7 @@ void Heatmap::run()
  *
  */
 
-float Heatmap::mapUnitsOf( float meters, QgsCoordinateReferenceSystem layerCrs )
+double Heatmap::mapUnitsOf( double meters, QgsCoordinateReferenceSystem layerCrs )
 {
   // Worker to transform metres input to mapunits
   QgsDistanceArea da;
@@ -318,7 +318,7 @@ float Heatmap::mapUnitsOf( float meters, QgsCoordinateReferenceSystem layerCrs )
   return meters / da.measureLine( QgsPoint( 0.0, 0.0 ), QgsPoint( 0.0, 1.0 ) );
 }
 
-int Heatmap::bufferSize( float radius, float cellsize )
+int Heatmap::bufferSize( double radius, double cellsize )
 {
   // Calculate the buffer size in pixels
